fix : migration complete
