+++ delivery-service/src/main/java/com/sobok/deliveryservice/delivery/controller/DeliveryFeignController.java
@@ -5,13 +5,16 @@ import com.sobok.deliveryservice.common.exception.CustomException;
 import com.sobok.deliveryservice.delivery.dto.info.AuthRiderInfoResDto;
 import com.sobok.deliveryservice.delivery.dto.payment.DeliveryRegisterDto;
 import com.sobok.deliveryservice.delivery.dto.payment.RiderPaymentInfoResDto;
+import com.sobok.deliveryservice.delivery.dto.request.RiderReqDto;
 import com.sobok.deliveryservice.delivery.dto.response.ByPhoneResDto;
 import com.sobok.deliveryservice.delivery.dto.response.DeliveryResDto;
 import com.sobok.deliveryservice.delivery.dto.response.RiderInfoResDto;
+import com.sobok.deliveryservice.delivery.dto.response.RiderResDto;
 import com.sobok.deliveryservice.delivery.entity.Delivery;
 import com.sobok.deliveryservice.delivery.repository.DeliveryRepository;
 import com.sobok.deliveryservice.delivery.repository.RiderRepository;
 import com.sobok.deliveryservice.delivery.service.DeliveryService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -30,6 +33,12 @@ public class DeliveryFeignController {
     private final RiderRepository riderRepository;
     private final DeliveryRepository deliveryRepository;
 
+    @PostMapping("/signup")
+    public ResponseEntity<RiderResDto> signup(@RequestBody @Valid RiderReqDto dto) {
+        log.info("rider signup 요청 들어옴");
+        RiderResDto response = deliveryService.riderCreate(dto);
+        return ResponseEntity.ok(response);
+    }
 
     @PostMapping("/findByPhoneNumber")
     public ResponseEntity<?> getUser(@RequestBody String phoneNumber) {
